fix: Move proposal description to separate component and fix loading flashes.
