+++ src/pages/Guilds/Proposal.tsx
@@ -1,10 +1,8 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { FiArrowLeft } from 'react-icons/fi';
 import styled from 'styled-components';
 import useEtherSWR from 'ether-swr';
 import { useParams } from 'react-router-dom';
-import contentHash from 'content-hash';
-import Markdown from 'markdown-to-jsx';
 import Skeleton from 'react-loading-skeleton';
 
 import { IconButton } from '../../components/Guilds/common/Button';
@@ -14,9 +12,8 @@ import ProposalVoteCard from '../../components/Guilds/ProposalSidebar/ProposalVo
 import ProposalStatus from '../../components/Guilds/ProposalStatus';
 import ProposalActionsCard from '../../components/Guilds/ProposalActionsCard';
 import UnstyledLink from '../../components/Guilds/common/UnstyledLink';
-import useIPFSFile from '../../hooks/Guilds/ipfs/useIPFSFile';
-import { ProposalMetadata } from '../../types/types.guilds';
 import AddressButton from '../../components/Guilds/AddressButton';
+import ProposalDescription from '../../components/Guilds/ProposalDescription';
 
 const PageContainer = styled(Box)`
   display: grid;
@@ -56,13 +53,6 @@ const PageTitle = styled.h3`
   margin: 1rem 0;
 `;
 
-const ProposalDescription = styled.div`
-  margin: 1.5rem 0;
-  line-height: 1.5;
-  font-size: 16px;
-  text-align: justify;
-`;
-
 const StyledIconButton = styled(IconButton)`
   padding: 0;
   margin-top: 5px;
@@ -82,42 +72,6 @@ const HeaderTopRow = styled(Box)`
   align-items: center;
   margin-bottom: 1rem;
 `;
-
-//TODO: separate as a different SubComponent
-const Description = () => {
-  const { guild_id: guildId, proposal_id: proposalId } = useParams<{
-    chain_name: string;
-    guild_id?: string;
-    proposal_id?: string;
-  }>();
-  const {
-    data: proposal,
-    error,
-    isValidating,
-  } = useEtherSWR([guildId, 'getProposal', proposalId]);
-
-  const decodedContentHash = useMemo(() => {
-    try {
-      return contentHash.decode(proposal.contentHash);
-    } catch (e) {
-      console.error(e);
-      return null;
-    }
-  }, [proposal]);
-  const metadata = useIPFSFile<ProposalMetadata>(decodedContentHash);
-
-  return (
-    <>
-      {!isValidating && !error && metadata?.description && (
-        <ProposalDescription>
-          <Markdown>{metadata?.description}</Markdown>
-        </ProposalDescription>
-      )}
-      {error && !isValidating && <div>Errror IFPS</div>}
-      {!error && isValidating && <div>loading</div>}
-    </>
-  );
-};
 
 const ProposalPage: React.FC = () => {
   const {
@@ -161,7 +115,7 @@ const ProposalPage: React.FC = () => {
 
         <AddressButton address={proposal?.creator} />
 
-        <Description />
+        <ProposalDescription />
 
         <ProposalActionsWrapper>
           <ProposalActionsCard />
